+++ start/src/App.vue
@@ -1,7 +1,14 @@
 <script lang="ts" setup></script>
 
 <template>
-  <f-button font-color="#fff" color="#ff0de5" round>自定义颜色</f-button>
+  <f-tooltip
+    content="右边的提示信息"
+    position="right"
+    background="red"
+    font-color="blue"
+  >
+    <f-tag type="warning">信息</f-tag>
+  </f-tooltip>
 </template>
 
-<style lang="scss" scoped></style>
+<style scoped></style>